APIO-104 Move PathIdentifierMapper to uri.mapper package
