+++ modules/apps/foundation/apio-architect/apio-architect-api/bnd.bnd
@@ -14,7 +14,6 @@ Export-Package:\
 	com.liferay.apio.architect.exception,\
 	com.liferay.apio.architect.function,\
 	com.liferay.apio.architect.functional,\
-	com.liferay.apio.architect.identifier.mapper,\
 	com.liferay.apio.architect.language,\
 	com.liferay.apio.architect.list,\
 	com.liferay.apio.architect.logger,\
@@ -31,4 +30,5 @@ Export-Package:\
 	com.liferay.apio.architect.single.model,\
 	com.liferay.apio.architect.supplier,\
 	com.liferay.apio.architect.uri,\
+	com.liferay.apio.architect.uri.mapper,\
 	com.liferay.apio.architect.url
