+++ modules/apps/foundation/apio-architect/apio-architect-api/src/main/resources/com/liferay/apio/architect/uri/mapper/packageinfo
