+++ modules/apps/foundation/apio-architect/apio-architect-application/src/main/java/com/liferay/apio/architect/application/internal/uri/mapper/PathLongIdentifierMapper.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.apio.architect.application.internal.identifier.mapper;
+package com.liferay.apio.architect.application.internal.uri.mapper;
 
 import com.liferay.apio.architect.error.ApioDeveloperError.UnresolvableURI;
 import com.liferay.apio.architect.functional.Try;
-import com.liferay.apio.architect.identifier.mapper.PathIdentifierMapper;
 import com.liferay.apio.architect.uri.Path;
+import com.liferay.apio.architect.uri.mapper.PathIdentifierMapper;
 import com.liferay.apio.architect.wiring.osgi.manager.representable.NameManager;
 
 import java.util.Optional;
