+++ modules/apps/foundation/apio-architect/apio-architect-wiring-osgi-api/src/main/java/com/liferay/apio/architect/wiring/osgi/manager/PathIdentifierMapperManager.java
@@ -31,8 +31,8 @@ public interface PathIdentifierMapperManager {
 	/**
 	 * Converts a {@code Path} to its equivalent identifier of type {@code T},
 	 * if a valid {@link
-	 * com.liferay.apio.architect.identifier.mapper.PathIdentifierMapper} can be
-	 * found. Returns {@code Optional#empty()} otherwise.
+	 * com.liferay.apio.architect.uri.mapper.PathIdentifierMapper} can be found.
+	 * Returns {@code Optional#empty()} otherwise.
 	 *
 	 * @param  clazz the class of the desired identifier
 	 * @param  path the {@code Path}
